Fix integer fast path overflows and compound assignment spec divergences (#2497)
* Fix int32 overflow in integer fast paths at the int.MinValue/MaxValue boundary

Integer-tagged arithmetic fast paths performed 32-bit operations that wrap
or raise hardware overflow exceptions at the int32 boundary:

- -((1<<31)|0) wrapped to -2147483648 instead of 2147483648
- ((1<<31)|0) / -1 and % -1 threw "Arithmetic operation resulted in an
  overflow" (CLR ArithmeticException escaping to the caller)
- 2147483647++ wrapped to -2147483648; ((1<<31)|0)-- and -= 1 wrapped
  to 2147483647

Widen to 64-bit math before the operation, matching the existing pattern
already used by the +, - and * binary operator fast paths.


* Fix compound assignment operators to match binary operator semantics

Compound assignments computed their results with simplified logic that
diverged from the corresponding binary operators:

- `u *= 2` and `u /= 2` with undefined operand produced undefined
  instead of NaN (the *= path short-circuited on undefined, the /= path
  passed raw operands into DivideComplex's undefined guard)
- `x **= Infinity` with |x| == 1 produced 1 via raw Math.Pow instead of
  NaN; ** and **= now share the spec-compliant Exponentiate helper
- compound bitwise/shift operators (&=, |=, ^=, <<=, >>=, >>>=) threw
  "Cannot convert a BigInt value to a number" for BigInt operands;
  they now share EvaluateBitwiseOperation with the binary forms,
  including the BigInt mixing TypeError and >>>= rejection
- -= and *= now apply ToNumeric before branching, producing the same
  "Cannot mix BigInt" TypeError as binary - and * for mixed operands


---------

+++ Jint/Runtime/Interpreter/Expressions/JintExpression.cs
@@ -167,7 +167,8 @@ internal abstract class JintExpression
             }
             else
             {
-                var modulo = leftInteger % rightInteger;
+                // long math: int.MinValue % -1 raises a hardware overflow in 32-bit division
+                var modulo = (long) leftInteger % rightInteger;
                 if (modulo == 0 && leftInteger < 0)
                 {
                     result = JsNumber.NegativeZero;
@@ -278,9 +279,10 @@ internal abstract class JintExpression
             return lN > 0 ? double.PositiveInfinity : double.NegativeInfinity;
         }
 
-        if (lN % rN == 0 && (lN != 0 || rN > 0))
+        // long math: int.MinValue / -1 (and % -1) raises a hardware overflow in 32-bit division
+        if ((long) lN % rN == 0 && (lN != 0 || rN > 0))
         {
-            return JsNumber.Create(lN / rN);
+            return JsNumber.Create((long) lN / rN);
         }
 
         return (double) lN / rN;
